pyproject.toml
[project]
name = "financial-market-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "pyarrow>=24.0.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "yfinance>=1.3.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "nbconvert>=7.17.1",
    "notebook>=7.5.5",
]
